backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic-settings==2.5.2
pydantic[email]==2.9.2
redis[asyncio]==5.1.0
celery[redis]==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
shapely==2.0.6
Pillow==10.4.0
firebase-admin==6.5.0
minio==7.2.9
python-multipart==0.0.12
websockets==13.1
aiofiles==24.1.0
geoalchemy2==0.15.2
pytest==8.3.3
pytest-asyncio==0.24.0
prometheus-client==0.21.0
starlette-prometheus==0.10.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
packaging>=24.0
uvloop==0.21.0
httptools==0.6.4
